chore(package.json): add script section
What will happen when a user clone the repo:
+ he will run "npm install"  to install dependencies
+ he will then check for script section…
+ …to see nothin can be run!

Webpack is known by the majority of node.js dev.
But:
+ not everyone knows webpack: an user that does not know webpack have
to search in the webpack site documentation about how to run scripts.
Taking a repo, installing, and runing it has to be the easier and most
comfortable possible.
+ an another-language dev will be strugling a lot about runing this repo

How script runs?
+ a script section is added in package.json
+ user will do (with npm): "npm run <script>"
+ there is a convention: "npm run start" has to run the basic command,
"npm run build" build the application, "npm run dev" is used in
development environement.