chore(webpack): Add local server
In order to be the most comfortable possible for the user,
a local server has been added, that will serve the "dist/" folder.
When typing "npm run serve", a server will be launched at
http://localhost:9000, along with webpack building the app.
Notice "dev" script is a shortcut of "npm run serve", a.k.a
"npm run dev" will launch "npm run serve".
That because serving a local server in a development environment
is comfortable for the user.